SurfaceData is now in surface.h. Namespace SessionData added.
