+++ session_data.cpp
@@ -1,67 +1,7 @@
 #include "session_data.h"
-#include "mapdata.h"
 #include <algorithm>
 
 using std::vector;
-
-//---------------------------------------------------------
-//---------------- SurfaceData ----------------------------
-//---------------------------------------------------------
-
-bool SurfaceData::Init(const str locationName) {
-    // open map, read it
-    // asign map pointers
-    // init factories
-
-    MapScanner scanner;
-    MapData data = scanner.getMap(locationName);
-    surfaceMatrix = data.surfaceMatrix;
-    mapHeight = data.mapHeight;
-    mapWidth = data.mapWidth;
-    GameData::setHeight(mapHeight);
-    GameData::setWidth(mapWidth);
-}
-
-int SurfaceData::getWidth() {
-    return mapWidth;
-}
-
-int SurfaceData::getHeight() {
-    return mapHeight;
-}
-
-short SurfaceData::getSurface(coord& point) {
-    if ((point.x <= 0 || point.x >= mapWidth) || (point.y <= 0 || point.y >= mapHeight))
-        return -1;
-    return surfaceMatrix[point.x][point.y];
-}
-
-bool SurfaceData::isValidCoord(coord& point) {
-    if ((point.x < 0 || point.x > mapWidth) || (point.y < 0 || point.y > mapHeight))
-        return false;
-    return true;
-}
-
-bool SurfaceData::isValidRadius(coord& point, short radius) {
-    if (
-        (point.x - radius <= 0) || 
-        (point.x + radius >= mapWidth) || 
-        (point.y + radius >= mapHeight) ||
-        (point.y - radius <= 0) ||
-        (!isValidCoord(point))
-    ) return false;
-
-    bool flag = false;
-    coord temp;
-    for (int i = point.x - radius; i < point.x + radius; i++) {
-        for (int j = point.y - radius; j < point.y + radius; j++) {
-            temp.x = i; temp.y = j;
-            if (getSurface(temp)) flag = true;
-        }
-    }
-    
-    return flag;
-}
 
 //---------------------------------------------------------
 //----------------- EventsData ----------------------------
@@ -70,10 +10,11 @@ bool SurfaceData::isValidRadius(coord& point, short radius) {
 bool EventsData::Init() {
     EventFactory eFactory;
     str path = "tests/events";
-    eFactory.InitAll(path, currentEventList); //hardcoded just for debugging
-    //using unique_ptr for two-dim array isn't a good idea though
+    eFactory.InitAll(path, currentEventList); // hardcoded just for debugging
+    // using unique_ptr for two-dim array isn't a good idea though
     // std::unique_ptr<Event[][]> eventMatrix (nullptr); 
-    PulverizeEvents(currentEventList);
+    PulverizeEvents(currentEventList); 
+    return true;
 }
 
 Event& EventsData::getEvent(const str key) {
@@ -93,8 +34,8 @@ add event to map
     for (auto i : list) {
         Event event = i.second;
         
-        int width = GameData::getWidth();
-        int height = GameData::getHeight();
+        int width = GameData::mapWidth;
+        int height = GameData::mapHeight;
 
         coord eventCenter = event.GetCoord();
         if (eventCenter.x > width || eventCenter.y > height) throw "coordinates are out of range";
@@ -140,28 +81,9 @@ bool ArtifactsData::Init() {
     ArtifactFactory aFactory;
     str path = "tests/artifacts";
     aFactory.InitAll(path, currentArtifactsList);
+    return true;
 }
 
 Artifact& ArtifactsData::getArtifact(const str key) {
     return currentArtifactsList.at(key);
-}
-
-//---------------------------------------------------------
-//------------------- GameData ----------------------------
-//---------------------------------------------------------
-
-short GameData::getHeight() {
-    return mapHeight;
-}
-
-short GameData::getWidth() {
-    return mapWidth;
-}
-
-void GameData::setHeight(short _height) {
-    mapHeight = _height;
-}
-
-void GameData::setWidth(short _width) {
-    mapWidth = _width;
 }
