+++ session_data.h
@@ -31,24 +31,6 @@ private:
     static short mapWidth;    
 };
 
-class SurfaceData {
-public:
-    bool Init(const str);
-    int getWidth();
-    int getHeight();
-    bool **GetMiniMap(); // coming soon
-    short getSurface(coord&);
-    Surface& getSurface(str);   
-    bool isValidRadius(coord&, short);
-    bool isValidCoord(coord&);
-private:
-    int mapWidth, mapHeight;
-    short** surfaceMatrix;
-    std::unordered_map<str, Surface> currentSurfaceList; // currently useless
-    // std::unordered_map<str, Surface> globalSurfaceList; //will be added in future
-
-};
-
 class EventsData {
 public:
     bool Init();
